Fixed Registration to properly persist in database. Allow users to cancel their registration.

+++ src/main/java/com/springfield/springboot/controller/ModuleController.java
@@ -26,8 +26,10 @@ public class ModuleController {
     @GetMapping("/modules")
     public String viewUserModules(Model model, HttpSession session)
             throws UserNotFoundException, ModuleNotFoundException {
-        @SuppressWarnings("unchecked")
         Long userID = (Long) session.getAttribute("CURRENT_USER");
+        if (userID == null)
+            return "redirect:logout?error=true";
+
         User user = userRepository.findById(userID)
                 .orElseThrow(() -> new UserNotFoundException(userID));
         model.addAttribute("user", user);
@@ -46,8 +48,9 @@ public class ModuleController {
     @GetMapping("/modules/all")
     public String viewAllModules(Model model, HttpSession session)
             throws UserNotFoundException, ModuleNotFoundException {
-        @SuppressWarnings("unchecked")
         Long userID = (Long) session.getAttribute("CURRENT_USER");
+        if (userID == null)
+            return "redirect:logout?error=true";
         if (session.getAttribute("userModules")==null) return "redirect:modules";
         List<Module> modules = moduleRepository.findAll();
         model.addAttribute("modules", modules);
@@ -55,7 +58,10 @@ public class ModuleController {
     }
 
     @RequestMapping("/statistics/{moduleID}/{moduleName}")
-    public String viewStatistics(@PathVariable(value = "moduleID") long moduleID, @PathVariable(value = "moduleName") String moduleName, Model model) {
+    public String viewStatistics(@PathVariable(value = "moduleID") long moduleID, @PathVariable(value = "moduleName") String moduleName, Model model, HttpSession session) {
+        Long userID = (Long) session.getAttribute("CURRENT_USER");
+        if (userID == null)
+            return "redirect:logout?error=true";
         JSONObject data = new JSONObject();
         data.put("female",involvementRepository.countUsersInvolvedBySex(moduleID, 'F') );
         data.put("male", involvementRepository.countUsersInvolvedBySex(moduleID, 'M') );
@@ -67,6 +73,8 @@ public class ModuleController {
     @RequestMapping("/enrol/{moduleID}")
     public String dropModule(@PathVariable(value = "moduleID") Long moduleID, Model model, HttpSession session) throws UserNotFoundException, ModuleNotFoundException {
         Long userID = (Long) session.getAttribute("CURRENT_USER");
+        if (userID == null)
+            return "redirect:logout?error=true";
         Module module = moduleRepository.findById(moduleID)
                 .orElseThrow(() -> new ModuleNotFoundException(moduleID));
         if (module.getIsFinished() == 'N' && module.getMaxStudents() >= 0+involvementRepository.studentsEnrolled(moduleID)) {
@@ -80,6 +88,8 @@ public class ModuleController {
     @RequestMapping("/drop/{moduleID}")
     public String enrolModule(@PathVariable(value = "moduleID") Long moduleID, Model model, HttpSession session) throws UserNotFoundException, ModuleNotFoundException {
         Long userID = (Long) session.getAttribute("CURRENT_USER");
+        if (userID == null)
+            return "redirect:logout?error=true";
         Involvement involvement = involvementRepository.findById(new InvolvementID(userID, moduleID))
                 .orElseThrow(() -> new ModuleNotFoundException(moduleID));
         involvementRepository.delete(involvement);
